refactor(migration-diagnostics): align export error ownership
