+++ src/renderer/windows/migrationV2/exporters/index.ts
@@ -4,4 +4,5 @@
 
 export { DexieExporter, type ExportProgress } from './DexieExporter'
 export { LocalStorageExporter } from './LocalStorageExporter'
+export { MigrationExportWriteError } from './MigrationExportWriteError'
 export { ReduxExporter, type ReduxExportResult } from './ReduxExporter'
