+++ src/shared/data/migration/v2/__tests__/diagnostics.test.ts
@@ -1,39 +1,12 @@
 import { describe, expect, it } from 'vitest'
 
 import {
-  assertMigrationExportWriteSucceeded,
   getMigrationDiagnosticNoticeParts,
   type MigrationDiagnosticError,
   type MigrationDiagnosticNoticePart,
   type MigrationDiagnosticSavedResult,
-  MigrationExportWriteError,
   serializeMigrationDiagnosticError
 } from '../diagnostics'
-
-describe('assertMigrationExportWriteSucceeded', () => {
-  it('throws a typed renderer error without losing Main failure details', () => {
-    const failure = {
-      code: 'export_file_write_failed' as const,
-      origin: 'main' as const,
-      operation: 'write_export_file' as const,
-      targetPath: '/tmp/migration_temp/topics.json',
-      error: { name: 'Error', message: 'permission denied', code: 'EACCES' }
-    }
-
-    expect(() => assertMigrationExportWriteSucceeded({ ok: false, failure })).toThrowError(MigrationExportWriteError)
-
-    try {
-      assertMigrationExportWriteSucceeded({ ok: false, failure })
-    } catch (error) {
-      expect(error).toBeInstanceOf(MigrationExportWriteError)
-      expect((error as MigrationExportWriteError).failure).toBe(failure)
-    }
-  })
-
-  it('accepts a successful Main write result', () => {
-    expect(() => assertMigrationExportWriteSucceeded({ ok: true })).not.toThrow()
-  })
-})
 
 describe('getMigrationDiagnosticNoticeParts', () => {
   it.each<[MigrationDiagnosticSavedResult, readonly MigrationDiagnosticNoticePart[]]>([
