+++ src/shared/data/migration/v2/diagnostics.ts
@@ -80,22 +80,6 @@ export type MigrationExportWriteResult =
   | { readonly ok: true }
   | { readonly ok: false; readonly failure: MigrationDiagnosticFailure }
 
-export class MigrationExportWriteError extends Error {
-  readonly failure: MigrationDiagnosticFailure
-
-  constructor(failure: MigrationDiagnosticFailure) {
-    super(failure.error?.message ?? failure.code)
-    this.name = 'MigrationExportWriteError'
-    this.failure = failure
-  }
-}
-
-export function assertMigrationExportWriteSucceeded(
-  result: MigrationExportWriteResult
-): asserts result is { ok: true } {
-  if (!result.ok) throw new MigrationExportWriteError(result.failure)
-}
-
 export type MigrationDiagnosticSavedResult =
   | { readonly status: 'saved'; readonly logs: 'included'; readonly size: 'standard' | 'large' }
   | {
